pyproject.toml
[project]
name = "fauna"
version = "0.1.0"
description = "Budgerigar contact-call encoder/decoder with intermediate semantic tagging"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "librosa>=0.10.1",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "soundfile>=0.12.1",
    "einops>=0.7.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
